+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - [#373](http://github.com/Nelson-numerical-software/nelson/issues/373): `sign` builtin.
 
+- [#313](http://github.com/Nelson-numerical-software/nelson/issues/313): `atanh` builtin: inverse hyperbolic tangent.
+
 - `MException` comes default exception in Nelson.
 
 - `try, catch` extended to manage `MException`.